Se crea caso de uso para la entidad doctor pokemon
