+++ Back/applications/app-service/src/main/java/co/com/sofka/config/UseCasesConfig.java
@@ -3,11 +3,13 @@ package co.com.sofka.config;
 import co.com.sofka.model.consulta.gateways.ConsultaRepository;
 import co.com.sofka.model.consulta.entity.mascotapokemon.gateways.MascotaPokemonRepository;
 import co.com.sofka.model.consulta.entity.usuario.gateways.UsuarioRepository;
+import co.com.sofka.model.doctorpokemon.gateways.DoctorpokemonRepository;
+import co.com.sofka.usecase.atencion.doctorpokemon.CreateDoctorPokemonUseCase;
 import co.com.sofka.usecase.consulta.CreateConsultaUseCase;
 import co.com.sofka.usecase.consulta.ListarConsultaUseCase;
-import co.com.sofka.usecase.mascota.CreateMascotaUseCase;
-import co.com.sofka.usecase.mascota.ListarMascotaUseCase;
-import co.com.sofka.usecase.usuario.CreateUsuarioUseCase;
+import co.com.sofka.usecase.consulta.mascota.CreateMascotaUseCase;
+import co.com.sofka.usecase.consulta.mascota.ListarMascotaUseCase;
+import co.com.sofka.usecase.consulta.usuario.CreateUsuarioUseCase;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
@@ -37,5 +39,9 @@ public class UseCasesConfig {
 
         public ListarConsultaUseCase listar(ConsultaRepository consultaRepository){
                 return new ListarConsultaUseCase(consultaRepository);
+        }
+
+        public CreateDoctorPokemonUseCase creardoctor(DoctorpokemonRepository doctorpokemonRepository){
+                return new CreateDoctorPokemonUseCase(doctorpokemonRepository);
         }
 }
