+++ Back/domain/model/src/main/java/co/com/sofka/model/doctorpokemon/gateways/DoctorpokemonRepository.java
@@ -1,4 +1,7 @@
 package co.com.sofka.model.doctorpokemon.gateways;
 
+import co.com.sofka.model.doctorpokemon.Doctorpokemon;
+
 public interface DoctorpokemonRepository {
+    Doctorpokemon creardoctor(Doctorpokemon doctorpokemon);
 }
