+++ Back/domain/usecase/src/main/java/co/com/sofka/usecase/consulta/mascota/ListarMascotaUseCase.java
@@ -1,15 +1,16 @@
-package co.com.sofka.usecase.mascota;
+package co.com.sofka.usecase.consulta.mascota;
 
 import co.com.sofka.model.consulta.entity.mascotapokemon.MascotaPokemon;
 import co.com.sofka.model.consulta.entity.mascotapokemon.gateways.MascotaPokemonRepository;
 import lombok.RequiredArgsConstructor;
 
+import java.util.List;
+
 @RequiredArgsConstructor
-public class CreateMascotaUseCase {
+public class ListarMascotaUseCase {
     private final MascotaPokemonRepository mascotaPokemonRepository;
 
-    public MascotaPokemon crearMascota(MascotaPokemon mascotaPokemon){
-        return mascotaPokemonRepository.crearMascota(mascotaPokemon);
+    public List<MascotaPokemon> listarPokemones(){
+        return mascotaPokemonRepository.listarPokemones();
     }
-
 }
