+++ Back/infrastructure/entry-points/api-rest/src/main/java/co/com/sofka/api/apirest/consultaapirest/mascotaapirest/AgregarMascotaApiRest.java
@@ -1,7 +1,7 @@
 package co.com.sofka.api.apirest.consultaapirest.mascotaapirest;
 
 import co.com.sofka.model.consulta.entity.mascotapokemon.MascotaPokemon;
-import co.com.sofka.usecase.mascota.CreateMascotaUseCase;
+import co.com.sofka.usecase.consulta.mascota.CreateMascotaUseCase;
 import lombok.AllArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
