+++ Back/infrastructure/entry-points/api-rest/src/main/java/co/com/sofka/api/apirest/consultaapirest/usuarioapirest/AgregarUsuariosApiRest.java
@@ -3,7 +3,7 @@ package co.com.sofka.api.apirest.consultaapirest.usuarioapirest;
 import co.com.sofka.api.dtos.UsuarioDTO;
 import co.com.sofka.api.mapper.UsuarioMapper;
 import co.com.sofka.model.consulta.entity.usuario.Usuario;
-import co.com.sofka.usecase.usuario.CreateUsuarioUseCase;
+import co.com.sofka.usecase.consulta.usuario.CreateUsuarioUseCase;
 import lombok.AllArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
